requirements.txt
PyQt6>=6.5.0
pymupdf>=1.23.0
openai>=1.3.0
python-dotenv>=1.0.0
pygame>=2.5.0
pyinstaller